refactor: simplify ce version parsers
