+++ src/transport/index.ts
@@ -1,6 +1,4 @@
 export * from "./emitter";
 export * from "./receiver";
-export * as v1Headers from "./http/v1";
-export * as v03Headers from "./http/v03";
 export * as http from "./http/headers";
 export * from "./protocols";
